refactor: iOS 전용 API와 서비스를 별도로 분리
