+++ src/test/java/com/techfork/domain/auth/service/AuthServiceTest.java
@@ -2,16 +2,12 @@ package com.techfork.domain.auth.service;
 
 import com.techfork.auth.application.AuthConverter;
 import com.techfork.auth.application.AuthService;
-import com.techfork.auth.application.KakaoOAuthService;
 import com.techfork.auth.application.dto.DeveloperTokenResponse;
-import com.techfork.auth.application.dto.KakaoLoginResponse;
 import com.techfork.auth.application.dto.TokenRefreshResponse;
-import com.techfork.auth.application.dto.kakao.KakaoUserInfoResponse;
 import com.techfork.auth.domain.exception.AuthErrorCode;
 import com.techfork.useraccount.domain.User;
 import com.techfork.useraccount.domain.enums.Role;
 import com.techfork.useraccount.domain.enums.SocialType;
-import com.techfork.useraccount.domain.enums.UserStatus;
 import com.techfork.useraccount.infrastructure.UserRepository;
 import com.techfork.global.exception.GeneralException;
 import com.techfork.global.security.auth.service.RefreshTokenService;
@@ -19,7 +15,6 @@ import com.techfork.global.security.auth.service.UserAuthCacheService;
 import com.techfork.global.security.jwt.JwtDTO;
 import com.techfork.global.security.jwt.JwtProperties;
 import com.techfork.global.security.jwt.JwtUtil;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -59,9 +54,6 @@ class AuthServiceTest {
 
     @Mock
     private AuthConverter authConverter;
-
-    @Mock
-    private KakaoOAuthService kakaoOAuthService;
 
     @Mock
     private UserAuthCacheService userAuthCacheService;
@@ -303,99 +295,4 @@ class AuthServiceTest {
         verify(jwtUtil, never()).generateLongLivedAccessToken(anyLong(), any(Role.class));
     }
 
-    // ===== 카카오 로그인 테스트 =====
-
-    @Test
-    @DisplayName("카카오 로그인 성공 - 신규 회원 가입")
-    void kakaoLogin_Success_NewUser() {
-        // Given
-        String kakaoAccessToken = "kakao.access.token";
-        String socialId = "12345";
-        String email = "newuser@kakao.com";
-        String profileImageUrl = "test.png";
-
-        KakaoUserInfoResponse.Profile profile = new KakaoUserInfoResponse.Profile("https://example.com/profile.jpg");
-        KakaoUserInfoResponse.KakaoAccount kakaoAccount = new KakaoUserInfoResponse.KakaoAccount(email, profile);
-        KakaoUserInfoResponse kakaoUserInfo = new KakaoUserInfoResponse(12345L, kakaoAccount);
-
-        User newUser = User.createSocialUser(SocialType.KAKAO, socialId, email, profileImageUrl);
-        ReflectionTestUtils.setField(newUser, "id", userId);
-
-        JwtDTO tokens = JwtDTO.of(newAccessToken, newRefreshToken);
-        KakaoLoginResponse expectedResponse = KakaoLoginResponse.builder()
-                .accessToken(newAccessToken)
-                .userId(userId)
-                .isRegistered(false)
-                .build();
-
-        given(kakaoOAuthService.getUserInfo(kakaoAccessToken)).willReturn(kakaoUserInfo);
-        given(userRepository.findBySocialTypeAndSocialId(SocialType.KAKAO, socialId)).willReturn(Optional.empty());
-        given(userRepository.save(any(User.class))).willReturn(newUser);
-        given(jwtUtil.generateTokens(userId, Role.USER)).willReturn(tokens);
-        given(jwtProperties.getRefreshTokenExpiration()).willReturn(900000L);
-        given(authConverter.toKakaoLoginResponse(newAccessToken, newUser)).willReturn(expectedResponse);
-
-        // When
-        KakaoLoginResponse result = authService.kakaoLogin(kakaoAccessToken, response);
-
-        // Then
-        assertThat(result.accessToken()).isEqualTo(newAccessToken);
-        assertThat(result.userId()).isEqualTo(userId);
-        assertThat(result.isRegistered()).isFalse(); // 신규 가입이므로 PENDING 상태
-
-        verify(kakaoOAuthService).getUserInfo(kakaoAccessToken);
-        verify(userRepository).findBySocialTypeAndSocialId(SocialType.KAKAO, socialId);
-        verify(userRepository).save(any(User.class));
-        verify(jwtUtil).generateTokens(userId, Role.USER);
-        verify(refreshTokenService).saveRefreshToken(eq(userId), eq(newRefreshToken), anyLong());
-        verify(response).addHeader(eq("Set-Cookie"), anyString());
-        verify(authConverter).toKakaoLoginResponse(newAccessToken, newUser);
-    }
-
-    @Test
-    @DisplayName("카카오 로그인 성공 - 기존 회원 로그인")
-    void kakaoLogin_Success_ExistingUser() {
-        // Given
-        String kakaoAccessToken = "kakao.access.token";
-        String socialId = "12345";
-        String email = "existinguser@kakao.com";
-        String profileImageUrl = "test.png";
-
-        KakaoUserInfoResponse.Profile profile = new KakaoUserInfoResponse.Profile("https://example.com/profile.jpg");
-        KakaoUserInfoResponse.KakaoAccount kakaoAccount = new KakaoUserInfoResponse.KakaoAccount(email, profile);
-        KakaoUserInfoResponse kakaoUserInfo = new KakaoUserInfoResponse(12345L, kakaoAccount);
-
-        User existingUser = User.createSocialUser(SocialType.KAKAO, socialId, email, profileImageUrl);
-        ReflectionTestUtils.setField(existingUser, "id", userId);
-        ReflectionTestUtils.setField(existingUser, "status", UserStatus.ACTIVE);
-
-        JwtDTO tokens = JwtDTO.of(newAccessToken, newRefreshToken);
-        KakaoLoginResponse expectedResponse = KakaoLoginResponse.builder()
-                .accessToken(newAccessToken)
-                .userId(userId)
-                .isRegistered(true)
-                .build();
-
-        given(kakaoOAuthService.getUserInfo(kakaoAccessToken)).willReturn(kakaoUserInfo);
-        given(userRepository.findBySocialTypeAndSocialId(SocialType.KAKAO, socialId)).willReturn(Optional.of(existingUser));
-        given(jwtUtil.generateTokens(userId, Role.USER)).willReturn(tokens);
-        given(jwtProperties.getRefreshTokenExpiration()).willReturn(900000L);
-        given(authConverter.toKakaoLoginResponse(newAccessToken, existingUser)).willReturn(expectedResponse);
-
-        // When
-        KakaoLoginResponse result = authService.kakaoLogin(kakaoAccessToken, response);
-
-        // Then
-        assertThat(result.accessToken()).isEqualTo(newAccessToken);
-        assertThat(result.userId()).isEqualTo(userId);
-        assertThat(result.isRegistered()).isTrue(); // 기존 회원이므로 ACTIVE 상태
-
-        verify(kakaoOAuthService).getUserInfo(kakaoAccessToken);
-        verify(userRepository).findBySocialTypeAndSocialId(SocialType.KAKAO, socialId);
-        verify(userRepository, never()).save(any(User.class)); // 기존 회원이므로 save 호출 안됨
-        verify(jwtUtil).generateTokens(userId, Role.USER);
-        verify(refreshTokenService).saveRefreshToken(eq(userId), eq(newRefreshToken), anyLong());
-        verify(response).addHeader(eq("Set-Cookie"), anyString());
-        verify(authConverter).toKakaoLoginResponse(newAccessToken, existingUser);
-    }
 }
